+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### [Unreleased]
 
+### [0.3.0] - 2017-03-06
+
+#### Changed
+* new namespaces for examples [Prokki\Warlight2BotTemplate\Examples](examples/).
+* [StupidRandomBot.php](examples/StupidRandomBot/src/StupidRandomBot.php) moved to example directory.
+* Base classes for bots moved to root directory [/src](src/).
+
 ### [0.2.0] - 2017-03-03
 
 #### Added
@@ -21,7 +28,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 #### Changed
 * Bot engine outsourced to https://github.com/prokki/theaigames-bot-engine.
 * Commands updated to use Bot instead of `AIable` (and [Environment](src/Game/Environment.php))
-* Renamed class AI to [AIBot.php](src/Bot/AIBot.php), class RandomBot to [StupidRandomBot.php](src/Bot/StupidRandomBot.php)
+* Renamed class AI to [AIBot.php](src/AIBot.php), class RandomBot to [StupidRandomBot.php](examples/StupidRandomBot/src/StupidRandomBot.php)
 * Method `RegionArray::getOffsets()` does not use `getArrayCopy()` anymore.
 * Method to filter regions by owner moved from [Map](src/Game/Map.php) (method `getRegions()`).
 to [RegionArray](src/Game/RegionArray.php).